Commit final

+++ src/FormCuentaBancaria.vue
@@ -1,23 +1,25 @@
 <template>
-  <div id="form" v-if="seen">
-    <h1>Formulario</h1>
-    <div>
-      <label>Numero:</label>
-      <input type="text" id="numero" name="numero" v-bind:value="cuenta.Numero"/>   
-    </div>
-    <div>
-      <label>Propietario:</label>
-      <input type="text" id="propietario" name="propietario" v-bind:value="cuenta.Propietario"/>   
-    </div>
+  <div id="form" >
+    <div v-if="seen">
+      <h1>Formulario Cuenta Bancaria</h1>
+      <div>
+        <label>Numero:</label>
+        <input type="text" id="numero" name="numero" v-bind:value="cuenta.Numero"/>   
+      </div>
+      <div>
+        <label>Propietario:</label>
+        <input type="text" id="propietario" name="propietario" v-bind:value="cuenta.Propietario"/>   
+      </div>
 
-    <div>
-      <label>Saldo:</label>
-      <input type="number" id="saldo" name="saldo"v-bind:value="cuenta.Saldo" />   
-    </div>
-    
+      <div>
+        <label>Saldo:</label>
+        <input type="number" id="saldo" name="saldo"v-bind:value="cuenta.Saldo" />   
+      </div>
+      
 
-    <input type="button" id="submit" value="Enviar" v-on:click="enviar"/>
-    <input type="button" id="submit" value="Eliminar" v-on:click="eliminar"/>
+      <input type="button" id="submit" value="Enviar" v-on:click="enviar"/>
+      <input type="button" id="submit" value="Eliminar" v-on:click="eliminar" v-if="cuenta.Id !== -1"/>
+    </div>
   </div>
 </template>
 
@@ -56,17 +58,20 @@
     },
     eliminar: function(){
       this.seen = false;
-      axios.delete('http://10.0.2.2:62270/api/CuentaBancaria/' + this.cuenta.Id)
-      .then(response => {
-        EventBus.$emit("updateCuenta", this.cuenta);
-      })
+      if(this.cuenta.Id !== -1){
+        axios.delete('http://10.0.2.2:62270/api/CuentaBancaria/' + this.cuenta.Id)
+        .then(response => {
+          EventBus.$emit("updateCuenta", this.cuenta);
+        })  
+      }
+      
     }
   },
   created() {
     this.cuenta = this.$parent.cuenta;
-    }
-
   }
+
+}
 </script>
 
 <style>
